fix: update usage module documentation and extract shared utilities
- Update SUMMARY.md to document global analytics capabilities, new
  use cases (GetGlobalProviderUsageUseCase, GetGlobalModelDistributionUseCase),
  super admin controllers, and cross-organization analytics features

- Extract shared utility functions to usage.utils.ts:
  - validateDateRange: date range validation logic
  - processModelDistribution: model distribution processing with percentages
  - calculateProviderPercentages: provider usage percentage calculation

- Refactor four use cases to use shared utilities, eliminating ~100 lines
  of duplicate code across GetModelDistributionUseCase,
  GetGlobalModelDistributionUseCase, GetProviderUsageUseCase, and
  GetGlobalProviderUsageUseCase

Applied via @cursor push command
