+++ ayunis-core-backend/src/domain/usage/application/use-cases/get-global-provider-usage/get-global-provider-usage.use-case.ts
@@ -2,8 +2,10 @@ import { Injectable } from '@nestjs/common';
 import { GetGlobalProviderUsageQuery } from './get-global-provider-usage.query';
 import { UsageRepository } from '../../ports/usage.repository';
 import { ProviderUsage } from '../../../domain/provider-usage.entity';
-import { InvalidDateRangeError } from '../../usage.errors';
-import { UsageConstants } from '../../../domain/value-objects/usage.constants';
+import {
+  validateDateRange,
+  calculateProviderPercentages,
+} from '../../usage.utils';
 
 @Injectable()
 export class GetGlobalProviderUsageUseCase {
@@ -11,49 +13,12 @@ export class GetGlobalProviderUsageUseCase {
 
   async execute(query: GetGlobalProviderUsageQuery): Promise<ProviderUsage[]> {
     if (query.startDate && query.endDate) {
-      this.validateDateRange(query.startDate, query.endDate);
+      validateDateRange(query.startDate, query.endDate);
     }
 
     const providerUsage =
       await this.usageRepository.getGlobalProviderUsage(query);
 
-    const totalTokens = providerUsage.reduce(
-      (sum, provider) => sum + provider.tokens,
-      0,
-    );
-
-    return providerUsage.map((provider) => {
-      const percentage =
-        totalTokens > 0 ? (provider.tokens / totalTokens) * 100 : 0;
-
-      return new ProviderUsage({
-        provider: provider.provider,
-        tokens: provider.tokens,
-        requests: provider.requests,
-        percentage,
-        timeSeriesData: provider.timeSeriesData,
-      });
-    });
-  }
-
-  private validateDateRange(startDate: Date, endDate: Date): void {
-    if (startDate > endDate) {
-      throw new InvalidDateRangeError('Start date cannot be after end date');
-    }
-
-    const now = new Date();
-    if (startDate > now) {
-      throw new InvalidDateRangeError('Start date cannot be in the future');
-    }
-
-    const maxDays = UsageConstants.MAX_DATE_RANGE_DAYS;
-    const daysDiff = Math.ceil(
-      (endDate.getTime() - startDate.getTime()) / (1000 * 60 * 60 * 24),
-    );
-    if (daysDiff > maxDays) {
-      throw new InvalidDateRangeError(
-        `Date range cannot exceed ${maxDays} days`,
-      );
-    }
+    return calculateProviderPercentages(providerUsage);
   }
 }
